+++ docs/CHANGELOG.md
@@ -5,6 +5,14 @@ Todos los cambios notables a este proyecto estarán en este documento.
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 y este proyecto se adhiere a [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.5] - 2025-10-30
+
+### Added
+
+- Clase base (Accion)
+- Clases herencia (AccionLanzar, AccionMover, AccionPoner, AccionSalir)
+
+
 ## [0.4.4] - 2025-10-29
 
 ### Added
@@ -14,8 +22,7 @@ y este proyecto se adhiere a [Semantic Versioning](https://semver.org/spec/v2.0.
 
 ### Changed
 
-- ahora poner_ficha coloca del lado correspondiente
-
+- Ahora poner_ficha coloca del lado correspondiente
 
 ## [0.4.3] - 2025-10-28